chore: flatten response
