+++ src/helpers/constants.js
@@ -1,5 +1,5 @@
 const MARKET_ARBITRAGE_THRESHOLD = 1;
-const COIN_GECKO_CHEQ_COIN_ID = "cheqd-network";
+const COIN_GECKO_CHEQ_COIN_ID = 'cheqd-network';
 
 exports.COIN_GECKO_CHEQ_COIN_ID = COIN_GECKO_CHEQ_COIN_ID;
 exports.MARKET_ARBITRAGE_THRESHOLD = MARKET_ARBITRAGE_THRESHOLD;
