+++ src/routes/arbitrage.js
@@ -1,9 +1,9 @@
-const express = require("express");
+const express = require('express');
 const router = express.Router();
 const {
   pricesAndPossibleArbitrageOpportunites,
-} = require("../app/pricesAndPossibleArbitrageOpportunites");
-router.get("/", async (_, res, __) => {
+} = require('../app/pricesAndPossibleArbitrageOpportunites');
+router.get('/', async (_, res, __) => {
   const data = await pricesAndPossibleArbitrageOpportunites();
   res.json(data);
 });
